Flexible Taxonomies
Broke out the taxonomy code written for Posts bundle in a way that it is more
easily reusable. Both tags and categories for Posts were broken out into
ProxySourceTaxonomyDataProvider and ProxySourceTaxonomyIndexGenerator. The
custom implementations in Posts bundle were all removed (four classes in total)
and the container configuration was updated slightly to just the generic classes
from contrib.

This means it should be somewhat trivial to add new taxonomy concepts to any
data provider that provides a collection of ProxySource ojbects.

Not 100% certain I'm using the correct terms here. :) But I think it should work
just fine for now.

Closes #56